// src/app/models/inspection.ts
export class Inspection {
  inspection: boolean = true;
  ref_no: String = '';
  supplier: String = '';
  contact: String = '';
  mobile: String = '';
  landline: String = '';
  email: String = '';
  inspection_date: any = '';
  shipment_date: any = '';
  allow_date_change: boolean;
  product_name: String = '';
  product_id: Number;
  sub_category_id: Number;
  category_id: Number;
  unit: String = '';
  purchase_order_no: String = '';
  quantity: Number;
  ref_or_sku: String = '';
}